backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
djangorestframework==3.16.0
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
twilio==9.6.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
